utility files for tests moved in dedicated folder
